chatbot/requirements.txt
# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0

# Database Drivers
psycopg2-binary==2.9.9
pymongo==4.6.1
SQLAlchemy==2.0.23

# Redis
redis==5.0.1

# LangGraph and AI
langgraph==0.0.62
langchain==0.2.0
langchain-core==0.2.0

# HTTP Client for Local LLM
requests==2.31.0

# Cloud LLM Providers (Optional)
openai==1.3.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# NLP & Fast Entity Detection
spacy==3.7.2

# Security
Werkzeug==3.0.1

# Production Server (Optional)
gunicorn==21.2.0



